- Move theme/ under core/
