+++ src/features/settings/appearance/themes/appTheme/AppThemePreview.tsx
@@ -1,6 +1,6 @@
 import { styled } from "@linaria/react";
 import { AppThemeType } from "../../../../../services/db";
-import { getThemeByStyle } from "../../../../../theme/AppThemes";
+import { getThemeByStyle } from "../../../../../core/theme/AppThemes";
 import { useIsDark } from "../../../../../core/GlobalStyles";
 
 const Container = styled.div<{
